+++ frontend/src/scenes/web-analytics/WebAnalyticsModal.tsx
@@ -1,6 +1,9 @@
 import { useActions, useValues } from 'kea'
 import { DateFilter } from 'lib/components/DateFilter/DateFilter'
+import { IconOpenInNew } from 'lib/lemon-ui/icons'
+import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { LemonModal } from 'lib/lemon-ui/LemonModal'
+import { urls } from 'scenes/urls'
 import { webAnalyticsLogic } from 'scenes/web-analytics/webAnalyticsLogic'
 import { WebQuery } from 'scenes/web-analytics/WebAnalyticsTile'
 import { WebPropertyFilters } from 'scenes/web-analytics/WebPropertyFilters'
@@ -27,20 +30,38 @@ export const WebAnalyticsModal = (): JSX.Element | null => {
             fullScreen={false}
             closable={true}
         >
-            <div className="flex flex-row flex-wrap gap-2 mb-4">
-                <WebPropertyFilters
-                    setWebAnalyticsFilters={setWebAnalyticsFilters}
-                    webAnalyticsFilters={webAnalyticsFilters}
-                />
-                <DateFilter dateFrom={dateFrom} dateTo={dateTo} onChange={setDates} />
+            <div className="space-y-4">
+                <div className="flex flex-row flex-wrap gap-2">
+                    <WebPropertyFilters
+                        setWebAnalyticsFilters={setWebAnalyticsFilters}
+                        webAnalyticsFilters={webAnalyticsFilters}
+                    />
+                    <DateFilter dateFrom={dateFrom} dateTo={dateTo} onChange={setDates} />
+                </div>
+                <LemonModal.Content embedded>
+                    <WebQuery
+                        query={modal.query}
+                        insightProps={modal.insightProps}
+                        showIntervalSelect={modal.showIntervalSelect}
+                    />
+                </LemonModal.Content>
+                <div className="flex flex-row justify-end">
+                    {modal.canOpenInsight ? (
+                        <LemonButton
+                            to={urls.insightNew(
+                                { properties: webAnalyticsFilters, date_from: dateFrom, date_to: dateTo },
+                                null,
+                                modal.query
+                            )}
+                            icon={<IconOpenInNew />}
+                            size="small"
+                            type="secondary"
+                        >
+                            Open as new Insight
+                        </LemonButton>
+                    ) : null}
+                </div>
             </div>
-            <LemonModal.Content embedded>
-                <WebQuery
-                    query={modal.query}
-                    insightProps={modal.insightProps}
-                    showIntervalSelect={modal.showIntervalSelect}
-                />
-            </LemonModal.Content>
         </LemonModal>
     )
 }
